Use the classes that have been moved to haystack-commons, deleting them here in secret-detector.
