+++ secret-detector/src/main/java/com/expedia/www/haystack/pipes/secretDetector/SpringWiredS3ConfigFetcher.java
@@ -1,7 +1,8 @@
 package com.expedia.www.haystack.pipes.secretDetector;
 
 import com.amazonaws.services.s3.AmazonS3;
-import com.expedia.www.haystack.pipes.secretDetector.config.WhiteListConfig;
+import com.expedia.www.haystack.commons.secretDetector.S3ConfigFetcher;
+import com.expedia.www.haystack.commons.secretDetector.WhiteListConfig;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -12,7 +13,7 @@ public class SpringWiredS3ConfigFetcher extends S3ConfigFetcher {
     SpringWiredS3ConfigFetcher(Logger s3ConfigFetcherLogger,
                                WhiteListConfig whiteListConfig,
                                AmazonS3 amazonS3,
-                               Factory s3ConfigFetcherFactory) {
+                               S3ConfigFetcher.Factory s3ConfigFetcherFactory) {
         super(s3ConfigFetcherLogger, whiteListConfig, amazonS3, s3ConfigFetcherFactory);
     }
 }
