+++ secret-detector/src/main/java/com/expedia/www/haystack/pipes/secretDetector/config/WhiteListConfigurationProvider.java
@@ -16,6 +16,7 @@
  */
 package com.expedia.www.haystack.pipes.secretDetector.config;
 
+import com.expedia.www.haystack.commons.secretDetector.WhiteListConfig;
 import org.cfg4j.provider.ConfigurationProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
