+++ secret-detector/src/main/java/com/expedia/www/haystack/pipes/secretDetector/mains/ProtobufSpanToEmailInKafkaTransformer.java
@@ -17,18 +17,19 @@
 package com.expedia.www.haystack.pipes.secretDetector.mains;
 
 import com.expedia.open.tracing.Span;
+import com.expedia.www.haystack.commons.secretDetector.Detector;
 import com.expedia.www.haystack.pipes.commons.kafka.KafkaConfigurationProvider;
 import com.expedia.www.haystack.pipes.commons.kafka.KafkaStreamBuilder;
 import com.expedia.www.haystack.pipes.commons.kafka.KafkaStreamStarter;
 import com.expedia.www.haystack.pipes.commons.kafka.Main;
 import com.expedia.www.haystack.pipes.commons.serialization.SpanSerdeFactory;
-import com.expedia.www.haystack.pipes.secretDetector.Detector;
 import org.apache.kafka.common.serialization.Serde;
 import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.streams.kstream.KStream;
 import org.apache.kafka.streams.kstream.KStreamBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+
 import static com.expedia.www.haystack.pipes.secretDetector.Constants.APPLICATION;
 
 @Component
